CLI-330 - fixes the .npm directory permission error in 2 approaches:
(a) setup npm environment to point to our own .npm cache directory inside our .appcelerator directory. this has a side effect (which i think is good) that the user's globally environment can't mess up our own installed environment and dependencies. this only sets the npm config using environment variable of the appc process such that our subprocess (appc-cli) which point to this directory for any npm work that it does

(b) run the npm install for modules that need to be run on the machine (native ones for example) in serial fashion instead of parallel. since npm can potentially create conflict lock files, this should also prevent further issues.  Wilson and I were both able to reproduce the main issue and resolve with (a), but this is still a good practice as Tony pointed out so i went ahead and did this anyway.