Add Coverage and Support for NPM3
The module lookup for alloy and other sub-commands is now more flexible. As it was, hard coding it to ../node_modules/appc-cli-titanium/node_modules... breaks with npm3 where all modules are flat in the top level node_modules.

Also, only run the sub-commands when the "--help" flag isn't there. This lets us run the help through the cli itself.